Feature #21931 Display roster data table  - Implemented a logic to convert ajax to post.  - Displayed students data on student roster page.  - Applied styling on student roster page.  - Implemented a logic to filter data by "Section".

+++ models/AssessmentSession.php
@@ -379,8 +379,11 @@ class AssessmentSession extends BaseImasAssessmentSessions
     public static function deleteByUserId($userId)
     {
         $assessmentSessions = AssessmentSession::find()->where(['userid' => $userId])->all();
-        foreach ($assessmentSessions as $assessmentSession) {
-            $assessmentSession->delete();
+        if($assessmentSessions)
+        {
+            foreach ($assessmentSessions as $assessmentSession) {
+                $assessmentSession->delete();
+            }
         }
     }
